RavenDB-22491 - Move SizeDetails out of DatabaseRequestHandler
