+++ src/Raven.Server/Documents/DatabaseRequestHandler.cs
@@ -281,25 +281,5 @@ namespace Raven.Server.Documents
             }
         }
 
-        public class SizeDetails : IDynamicJson
-        {
-            public int ActualSize { get; set; }
-            public string HumaneActualSize { get; set; }
-            public int AllocatedSize { get; set; }
-            public string HumaneAllocatedSize { get; set; }
-            public bool IsCompressed { get; set; }
-
-            public virtual DynamicJsonValue ToJson()
-            {
-                return new DynamicJsonValue
-                {
-                    [nameof(ActualSize)] = ActualSize,
-                    [nameof(HumaneActualSize)] = HumaneActualSize,
-                    [nameof(AllocatedSize)] = AllocatedSize,
-                    [nameof(HumaneAllocatedSize)] = HumaneAllocatedSize,
-                    [nameof(IsCompressed)] = IsCompressed
-                };
-            }
-        }
     }
 }
